Created project folders, a homepage, and a webpage
I created the images folder and recipes folder to organize my files. I then created the index.html with a basic boilerplate and customized it to fit the recipes project. I also made the first recipe's webpage and added a link to it in the index.html. Finally, I downloaded a picture to the images folder and added it on the new webpage.